Change2 log4j2.2.0.3

+++ src/main/java/com/tuanpla/utils/config/PublicConfig.java
@@ -22,7 +22,7 @@ public abstract class PublicConfig {
 
     public static long AUTO_LOAD_LIB_TIME = 30 * 1000;
     public static long DEFAULT_EXPIRE_TIME = 5 * 1000;
-    public static long MAX_FILE_SIZE = 10 * 1024 * 1024;
+    public static int MAX_FILE_SIZE = 10 * 1024 * 1024;
 
     // FOR FINAL CONSTANT
     public static final String UTF_8 = StandardCharsets.UTF_8.name();
